feat(mobile-calendar): replace events tab with unified calendar tab
Adds the new Calendar tab screen and removes the old Events tab. The tab
slot in _layout.tsx is renamed events → calendar (icon stays Calendar),
and the two ActionSheet handlers (RSVP, check-in) are repointed at
/(tabs)/calendar. The events feature stack under [orgSlug]/events/ stays
intact — only the tab entry point moves.

The new screen renders SourceFilterChips + UnifiedCalendarFeed against
useUnifiedCalendar(orgId), follows the gradient header + content sheet
pattern, and includes the admin overflow menu (Create Event, Open in Web).
